+++ doc/CHANGELOG.md
@@ -9,7 +9,8 @@ Changelog
 * UPDATE and MESSAGE methods.
 * Full event support (SUBSCRIBE/NOTIFY).
 * Full PUBLISH support, using in-memory or external database.
-* Path support.
+* Path support, as client, proxy and registrar.
+* Support for headers in URIs.
 * UAS callback functions receive contextual metadata.
 * New options to customize supported extensions and to generate Require and Accept headers.
 * Use of any external store for registrar instead of in-memory built-in.